update: update file name and delete useless element
