--- components/layout/HeaderTest.vue
@@ -1,74 +1,0 @@
-<template>
-    <header class="PC-menu-box">
-        <div class="logo-and-search-section">
-            <div class="logo-box">
-                <a href="/"><img ref="" class="logo" src="@/assets/img/logo.png" alt="logo"></img></a>
-            </div>
-            <div class="input-section">
-                <el-input class="input" v-model="search" :suffix-icon="Search"></el-input>
-            </div>
-            <div class="contact-section">
-                <nuxt-link class="contact" :to="'/'">聯絡我們</nuxt-link>
-            </div>
-        </div>
-    </header>
-</template>
-<script lang="ts" setup>
-import { Search } from '@element-plus/icons-vue'
-
-const search = ref('')
-</script>
-<style lang="scss" scoped>
-.logo-and-search-section {
-    background-color: #FFFFFF;
-    display: flex;
-    padding-top: 2%;
-
-    .logo-box {
-        flex-grow: 4;
-        margin-left: 40%;
-
-        .logo {
-            width: 20rem;
-        }
-    }
-
-    .input-section {
-        flex-grow: 1;
-        display: flex;
-        align-self: self-end;
-        padding-bottom: 1%;
-        :deep(.el-input__wrapper) {
-            background-color: $main-color;
-            max-height: 1.5rem;
-            width: 10rem;
-            border-radius: 30px;
-        }
-
-        :deep(.el-input__suffix) {
-            background-color: #FFFFFF;
-            color: $main-color;
-            padding-right: 1rem;
-            max-height: 1rem;
-            width: 0.8rem;
-            border-radius: 10px;
-            cursor: pointer;
-        }
-    }
-
-    .contact-section {
-        flex-grow: 1;
-        padding-left: 2rem;
-        display: flex;
-        align-items: self-end;
-        padding-bottom: 1%;
-
-
-        .contact {
-            color: $main-color;
-            font-size: 1rem;
-        }
-    }
-}
-
-</style>
